Fix package names to be similar to org.cotton.base
