--- catowserAndroid/app/src/main/java/com/cotton/MainBrowserView.kt
@@ -1,35 +1,0 @@
-package com.cotton
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import com.browser.content.*
-import org.cotton.base.Site
-
-@Composable
-internal fun MainBrowserView(viewModel: MainBrowserViewModel, searchBarViewModel: SearchBarViewModel) {
-    Column(modifier = Modifier.fillMaxSize()) {
-        Row(modifier = Modifier
-            .height(viewModel.barHeight)
-            .fillMaxWidth()) {
-            Column(modifier = Modifier.weight(1f)) {
-                SearchBarView(searchBarViewModel)
-            }
-            Column(modifier = Modifier
-                .width(30.dp)
-                .fillMaxHeight(), horizontalAlignment = Alignment.End) {
-                TabsCountButton(viewModel.onOpenTabs, 0u)
-            }
-        }
-        Row(modifier = Modifier.fillMaxSize()) {
-            if (viewModel.matchesFound) {
-                SearchSuggestionsView()
-            } else {
-                val content = TabContentType.SiteContent(Site.opennetru)
-                BrowserContent(content)
-            }
-        }
-    } // column
-}
