--- catowserAndroid/app/src/main/java/com/cotton/MainBrowserViewModel.kt
@@ -1,10 +1,0 @@
-package com.cotton
-
-import androidx.compose.ui.unit.dp
-import androidx.lifecycle.ViewModel
-
-class MainBrowserViewModel: ViewModel() {
-    val onOpenTabs: () -> Unit = {}
-    val barHeight = 50.dp
-    val matchesFound: Boolean = false
-}
