--- catowserAndroid/app/src/main/java/com/cotton/ui/theme/Color.kt
@@ -1,8 +1,0 @@
-package com.cotton.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple200 = Color(0x36F5B070)
-val Purple500 = Color(0xFFF56C1D)
-val Purple700 = Color(0xFFD92923)
-val Teal200 = Color(0xFFE9E9E9)
