--- catowserAndroid/browser-content/src/main/java/com/browser/content/SearchBarViewModel.kt
@@ -1,10 +1,0 @@
-package com.browser.content
-
-import androidx.lifecycle.ViewModel
-
-class SearchBarViewModel: ViewModel() {
-    val searchText: String = ""
-    val placeholderText: String = "Search or enter address"
-    val onSearchTextChanged: (String) -> Unit = {}
-    val onClearClick: () -> Unit = {}
-}
