--- catowserAndroid/browser-content/src/main/java/com/browser/content/TabContentType.kt
@@ -1,11 +1,0 @@
-package com.browser.content
-
-import org.cotton.base.Site
-
-sealed class TabContentType {
-    class Blank(): TabContentType()
-    class TopSites(): TabContentType()
-    class Favorites(): TabContentType()
-    class Homepage(): TabContentType()
-    class SiteContent(val site: Site): TabContentType()
-}
