--- catowserAndroid/browser-content/src/main/java/com/browser/content/ui/theme/Color.kt
@@ -1,8 +1,0 @@
-package com.browser.content.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple200 = Color(0x36F5B070)
-val Purple500 = Color(0xFFF56C1D)
-val Purple700 = Color(0xFFD92923)
-val Teal200 = Color(0xFFE9E9E9)
