+++ catowserAndroid/browser-content/src/main/java/org/cotton/browser/content/SearchBarViewModel.kt
@@ -1,0 +1,10 @@
+package org.cotton.browser.content
+
+import androidx.lifecycle.ViewModel
+
+class SearchBarViewModel: ViewModel() {
+    val searchText: String = ""
+    val placeholderText: String = "Search or enter address"
+    val onSearchTextChanged: (String) -> Unit = {}
+    val onClearClick: () -> Unit = {}
+}
